server/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.2
django-filter==22.1
pytest-django==4.5.2
pytest==6.2.5
drf-yasg==1.20.0
debugpy==1.5.1